debugging loading of plugins over 0.4.1.2-alpha

+++ src/ubpf/vm/ebpf.h
@@ -26,7 +26,7 @@ struct ebpf_inst {
     uint8_t dst : 4;
     uint8_t src : 4;
     int16_t offset;
-    int64_t imm;
+    int32_t imm;
 };
 
 #define EBPF_CLS_MASK 0x07
